+++ src/lib/components/forms/checkbox/terra-checkbox.component.spec.ts
@@ -3,26 +3,17 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { DebugElement } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { TooltipDirective } from '../../tooltip/tooltip.directive';
-import { Router } from '@angular/router';
-import { MockRouter } from '../../../testing/mock-router';
+import { MockTooltipDirective } from '../../../testing/mock-tooltip.directive';
 import Spy = jasmine.Spy;
 
 describe('Component: TerraCheckboxComponent', () => {
     let component: TerraCheckboxComponent;
     let fixture: ComponentFixture<TerraCheckboxComponent>;
-    const router: MockRouter = new MockRouter();
 
     beforeEach(() => {
         TestBed.configureTestingModule({
-            declarations: [TooltipDirective, TerraCheckboxComponent],
-            imports: [FormsModule],
-            providers: [
-                {
-                    provide: Router,
-                    useValue: router
-                }
-            ]
+            declarations: [MockTooltipDirective, TerraCheckboxComponent],
+            imports: [FormsModule]
         });
     });