+++ src/lib/components/tables/data-table/table-row/table-row.component.spec.ts
@@ -1,15 +1,13 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { TableRowComponent } from './table-row.component';
 import { TerraDataTableComponent } from '../terra-data-table.component';
-import { TerraCheckboxComponent } from '../../../forms/checkbox/terra-checkbox.component';
 import { FormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
 import { TerraDataTableRowInterface } from '../interfaces/terra-data-table-row.interface';
 import { Component, DebugElement } from '@angular/core';
-import { TooltipDirective } from '../../../tooltip/tooltip.directive';
-import { Router } from '@angular/router';
-import { MockRouter } from '../../../../testing/mock-router';
 import { noop } from 'rxjs';
+import { MockTooltipDirective } from '../../../../testing/mock-tooltip.directive';
+import { MatCheckboxModule } from '@angular/material/checkbox';
 
 export const dataTableStub: Partial<TerraDataTableComponent<any, any>> = {
     rowClicked: noop,
@@ -34,18 +32,11 @@ describe('Component: TableRowComponent', () => {
     let fixture: ComponentFixture<HostComponent>;
     let component: TableRowComponent;
     let dataTable: TerraDataTableComponent<any, any>;
-    const router: MockRouter = new MockRouter();
 
     beforeEach(() => {
         TestBed.configureTestingModule({
-            declarations: [TooltipDirective, TableRowComponent, TerraCheckboxComponent, HostComponent],
-            imports: [FormsModule],
-            providers: [
-                {
-                    provide: Router,
-                    useValue: router
-                }
-            ]
+            declarations: [MockTooltipDirective, TableRowComponent, HostComponent],
+            imports: [FormsModule, MatCheckboxModule]
         });
     });
 
@@ -101,19 +92,19 @@ describe('Component: TableRowComponent', () => {
     });
 
     it('should toggle checkbox visibility', () => {
-        expect(fixture.debugElement.query(By.css('terra-checkbox'))).toBeTruthy();
+        expect(fixture.debugElement.query(By.css('mat-checkbox'))).toBeTruthy();
 
         dataTable.inputHasCheckboxes = false;
 
         fixture.detectChanges();
 
-        expect(fixture.debugElement.query(By.css('terra-checkbox'))).toBeFalsy();
+        expect(fixture.debugElement.query(By.css('mat-checkbox'))).toBeFalsy();
     });
 
     it('should call #_onRowCheckboxChange() when checkbox changes', () => {
         dataTable.inputHasCheckboxes = true;
 
-        let rowData: TerraDataTableRowInterface<any> = {
+        const rowData: TerraDataTableRowInterface<any> = {
             isActive: false,
             selected: false,
             disabled: false
@@ -125,7 +116,7 @@ describe('Component: TableRowComponent', () => {
 
         fixture.detectChanges();
 
-        let checkbox: DebugElement = fixture.debugElement.query(By.css('terra-checkbox'));
+        const checkbox: DebugElement = fixture.debugElement.query(By.css('mat-checkbox'));
 
         checkbox.triggerEventHandler('change', {});