+++ src/lib/components/tree/node-tree/data/terra-node-tree.config.ts
@@ -1,6 +1,5 @@
 import { TerraNodeInterface } from './terra-node.interface';
 import { isNullOrUndefined } from 'util';
-import { L10nTranslationService } from 'angular-l10n';
 import { Injectable } from '@angular/core';
 import { TerraNodeTreeHelper } from '../helper/terra-node-tree.helper';
 import { ObjectHelper } from '../../../../helpers/object.helper';
@@ -12,8 +11,6 @@ import { ObjectHelper } from '../../../../helpers/object.helper';
 export class TerraNodeTreeConfig<D> {
     protected _currentSelectedNode: TerraNodeInterface<D>;
     private _list: Array<TerraNodeInterface<D>> = [];
-
-    constructor(public translation: L10nTranslationService) {}
 
     /**
      * @description Adds a node.